requirements.txt
dj-database-url==0.4.1
Django==1.9.6
django-widget-tweaks==1.4.1
djangorestframework==3.3.3
gunicorn==19.5.0
psycopg2==2.6.1
whitenoise==3.1
